new: 코스 조회 API 연결, 코스 리스트 -> 코스 상세 연결

+++ app/src/main/java/com/footprint/footprint/domain/model/CourseInfoModel.kt
@@ -6,9 +6,9 @@ data class CourseInfoModel(
     val idx: Int = 1,
     var previewImageUrl: String = "",
     var title: String = "",
-    var distance: Int? = null,
+    var distance: Double? = null, // Int?
     var time: Int? = null,
     var tags: List<String> = listOf(),
-    var coords: List<List<LatLng>> = listOf(listOf()),
+    var coords: List<List<LatLng>> = listOf(listOf()), // List<List<LatLng>> = listOf(listOf())
     var description: String = ""
 )
